Restructured component directory.
