+++ src/app/about/page.tsx
@@ -1,8 +1,8 @@
 // Dependencies.
 import { Fragment } from "react"
-import Header from "@/components/design/Header/Header"
-import AboutContent from "@/components/design/About/AboutContent"
-import Footer from "@/components/design/Footer/Footer"
+import Header from "@/components/Layout/Header/Header"
+import AboutContent from "@/components/About/About"
+import Footer from "@/components/Layout/Footer/Footer"
 
 // Page.
 export default function About() {
