+++ src/app/page.tsx
@@ -1,8 +1,8 @@
 // Dependencies.
 import { Fragment } from "react"
-import Header from "@/components/design/Header/Header"
-import HomeContent from "@/components/design/Home/HomeContent"
-import Footer from "@/components/design/Footer/Footer"
+import Header from "@/components/Layout/Header/Header"
+import HomeContent from "@/components/Home/Home"
+import Footer from "@/components/Layout/Footer/Footer"
 
 // Page.
 export default function Home() {
