+++ src/components/About/Team.tsx
