+++ src/components/About/Values.tsx
